real_estate_chatbot/requirements.txt
langchain>=0.0.335
langchain-google-genai>=0.0.6
langgraph>=0.0.25
streamlit>=1.29.0
python-dotenv>=1.0.0
pillow>=10.0.0
google-api-python-client>=2.100.0
google-search-results>=2.4.2
pydantic>=2.4.0
typing-extensions>=4.8.0
google-generativeai>=0.3.0
